Recursive
I use Divide-and-Conquer method to implement recursive.
1. Check the last element first.
If the character bigger than the last one, return the first element directly.
2. Compare with the middle element
If bigger, divide the string into two part, and do recursive with the
upper half.
If non-bigger, do recursive with lower half.
3. Terminate
The recursive will terminate when the two bound meets.
It compares with the meeting point and decide return the point or the
next element.